Fix dashboard endpoint
The sql query was incorrect. The dashboard tasks were only sorted by the
date of project creation, they needed further sorting by column (within
each project) and row (within each column). The pagination has also
been fixed as previously we were using the task id however seen as
though the projects are sorted as they are, this means that the taskId
isn't gaurenteed to be in a specific order and thus relying on this for
pagination was incorrect, we now use LIMIT and OFFSET; this isn't very
efficent as it doesn't use indexes and even if it did they won't be
worthwhile as using OFFSET and LIMIT fetches the whole data up until the
OFFSET point. In future maybe look into using a different method.