Fix login bug when user doesn't belong to organisation
It's possible that a user doesn't to an organisation (to create an
organisation you first must create an account, then the account is
assigned the admin of the organisation). When a user without an
organisation logged in, when adding access controls, the user didn't
have any projects or tasks this caused the server to break as you can't
add empty sets to redis. We now don't add anything at all when there is
an 'empty set' and then when we query redis it just returns false which
represents an empty set