Update attribute access controls on create project
Previously we only updated the user that created the asset (e.g.
project/task/projectcolumn etc...) however we need to update any user
that was logged in that could access the asset. This goes hand in with a
DB DDL change - admins can access all tasks including private, before
we weren't storing admins in the user_project table, we were going to
program in a clause where we check to see if the user is admin first and
then check attribute access controls but this would add extra
complexity. Now we have a single point of truth.