Get current task detials on edit task endpoint
Previously certain current task details (projectId, projectColumnId,
projectRow) were sent with the api request, this done to try to save
having to query the db for these details when they will already be
avaiable on the frontend. This has been removed, all of these details
are now fetched in one query on load for security, we were also getting
the valid project column ids in a query so it made sense to put all of
this into one query.

We were also getting the max rows for the old column and the new column
inside the transactions when moving columns, we do this in the
aforementioned query now.