Refactor backend
It was getting too messy to carry on in the same style. I realised that
i was structuring the files around the order in which i was creating
things, so i was doing the get requests, then the posts etc... this is
not the best way to organise code. Feels much better to organise into
the resources them selves and then split these into routes (app.use()
but using the router to break things down into 'miniapps'), controller
(the endpoints themselves) and then services (the db calls - this
finally split up the god awful db file).