Fix access controll bug
If a user didn't have a resource assigned and a new one was added, that
the user could access, this resource was not added to the access control
. addRemoveMultipleAttributeAccess skipped adding the resource if the
set wasn't present in the cachw. This was to make sure that no orphan sets were
generated in the cache. To keep this strictness in place a 'force' flag
has been added so that in certain instances we can add the set and the
resource at the same time