Add app structure
This includes the 'twitter' style modal popup for tasks which has it's
own unique url even - if the user clicks through from a link in the app
then the task opens as a modal whilst retaining the state of the app
however if the task is opened via a link then it will open as a modal on
a blank page.

Notice frontend/jsconfig.json here - this is needed when using an alias
so the 'to definition' works.