Fix get endpoint bug
When we changed the way we handled returning data from the api in
(#63398d) it broke the get endpoints. We have normalised the way we pass
data from the createEndpoint functions for get requests here, and we
have also made it so that if we don't need to output a success message
on the frontend from the getData composable then we don't.