Fix sidebars disapearing on scroll bug
The main sidebar should be always be in the main viewport  no matter if
the main content area can scroll. When in the project area the project
sidebar should also stay in the main view port too regardless of the
scroll. This has taken the scroll of the body and added it to the main
element.

Avoided using fixed elements here as this keeps the layout predictable