Add drag to tasks
This is the first phasae of adding drag and drop. It includes dragging a
task, adding the infrastructure for adding placeholders (the empty task
that will appear under the cursor to show where the task will be moved
to when dropped), and adding a placeholder on drag start, and removing
the placeholder on dragend.